Refactoring home sections in other files
