+++ src/pagePartials/home/BadgeCreator.tsx
@@ -1,0 +1,82 @@
+import TaskAltIcon from '@mui/icons-material/TaskAlt'
+import { Box, Typography } from '@mui/material'
+import { ButtonV2, colors } from 'thebadge-ui-library'
+
+export default function BadgeCreator() {
+  return (
+    <Box
+      alignItems={'center'}
+      display={'flex'}
+      flex={1}
+      flexDirection={'column'}
+      justifyContent={'space-between'}
+      paddingX={5}
+      paddingY={5}
+    >
+      <Typography
+        color={colors.pink}
+        fontSize={'30px !important'}
+        fontWeight={700}
+        lineHeight={'30px'}
+        mb={4}
+      >
+        Badge Creator
+      </Typography>
+      <Box alignItems={'center'} display={'flex'} flexDirection={'row'} justifyContent={'center'}>
+        <TaskAltIcon color={'pink'} sx={{ mr: 2 }} />
+        <Typography
+          component={'div'}
+          fontSize={'12px !important'}
+          fontWeight={600}
+          lineHeight={'16px'}
+        >
+          You can design and publish custom badges that other users can apply for to verify
+          real-world information.
+        </Typography>
+      </Box>
+      <Box
+        alignItems={'center'}
+        display={'flex'}
+        flexDirection={'row'}
+        justifyContent={'center'}
+        mt={2}
+      >
+        <TaskAltIcon color={'pink'} sx={{ mr: 2 }} />
+        <Typography
+          component={'div'}
+          fontSize={'12px !important'}
+          fontWeight={600}
+          lineHeight={'16px'}
+        >
+          Set the price for the badges you create and receive a share of the fee when users apply
+          for and receive them.
+        </Typography>
+      </Box>
+      <Typography
+        component={'span'}
+        fontSize={'12px !important'}
+        fontWeight={700}
+        lineHeight={'14px'}
+        mt={2}
+      >
+        Learn more
+      </Typography>
+      <ButtonV2
+        backgroundColor={colors.pink}
+        fontColor={colors.white}
+        sx={{
+          mt: 4,
+          borderRadius: '10px',
+          fontSize: '12px !important',
+          padding: '0.5rem 1rem !important',
+          height: 'fit-content !important',
+          lineHeight: '14px',
+          fontWeight: 700,
+          boxShadow: 'none',
+        }}
+      >
+        BECOME A CREATOR
+      </ButtonV2>
+    </Box>
+  )
+}
